Fixed currencyRegex, made setCurrencySymbol replace old symbol in text.
currencyRegex was not correctly registering $ as a symbol - in regex the
dollar symbol is reserved. Added escape symbol before it.

startsWithSymbol(String) was changed to startsWithSymbol(String,
CurrencySymbol) to facilitate this. It was also changed to public static
as it could be useful for users.

public static replaceOldSymbol(String input, CurrencySymbol oldSymbol,
CurrencySymbol newSymbol) was also added to facilitate the change.